requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
anthropic==0.84.0
python-dotenv==1.2.2
httpx==0.28.1
opencv-python==4.13.0.92
mediapipe==0.10.14
numpy==2.4.3
PyJWT==2.12.1
pypdf==6.8.0
python-docx==1.2.0
pydantic==2.12.5
starlette==0.52.1
python-multipart==0.0.20
